pyproject.toml
[tool.poetry]
name = "nonebot-plugin-randpic"
version = "1.0.0"
description = "一个通过指令随机发送你存储的图片的插件"
authors = ["HuParry <hu_parry@qq.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "nonebot_plugin_randpic"}]
homepage = "https://github.com/HuParry/nonebot-plugin-randpic"
repository = "https://github.com/HuParry/nonebot-plugin-randpic"


[tool.poetry.dependencies]
python = "^3.8"
nonebot2 = {version = "^2.0.0", extras = ["fastapi"] }
nonebot-adapter-onebot = "^2.1.5"
nonebot-plugin-localstore = "^0.6.0"
aiosqlite = "^0.13.0"
httpx = ">=0.20.0,<1.0.0"
alibabacloud_oss_v2 = ">=1.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
